refactor: ViewModel 프로토콜 파일 분리 #77
- Common 폴더 정리
